updated float to int to fix mismatch

+++ plugins/action/client_health_info.py
@@ -29,7 +29,7 @@ argument_spec = dnac_argument_spec()
 # Add arguments specific for this module
 argument_spec.update(
     dict(
-        timestamp=dict(type="int"),
+        timestamp=dict(type="float"),
         headers=dict(type="dict"),
     )
 )
